Landing Page Added
Landing page created, and some simple react-spring animations implemented.
