+++ src/Footer.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const Footer = () => {
+    return (
+    <footer>
+        <p>
+        A project by <a href="http://www.danielcranney.com" target="_blank">Daniel Cranney</a>
+        </p>
+    </footer>
+    )
+}
+
+export default Footer;
