+++ src/Header.js
@@ -2,17 +2,17 @@ import React from 'react'
 
 const Header = () => {
     return (
-    <div className="header">
+    <header>
       <div>
         <a href="index.html"><img src="/img/logo.svg" width="200px" height="71px" alt="Temporary Logo" /></a>
       </div>
       <nav>
         <a href="http://">About</a>
         <a href="http://">Donate</a>
-        <a href="http://">Visit BBFC</a>
+        <a href="https://www.bbfc.co.uk/" target="_blank">Visit BBFC</a>
         <a href="http://">Contact</a>
       </nav>
-    </div>
+    </header>
     )
 }
